Add support for using `ConsensusParametersV2`

+++ fuel-vm/src/util.rs
@@ -380,7 +380,7 @@ pub mod test_helpers {
             self.consensus_params.block_gas_limit()
         }
 
-        pub fn get_block_transaction_size_limit(&self) -> u32 {
+        pub fn get_block_transaction_size_limit(&self) -> u64 {
             self.consensus_params.block_transaction_size_limit()
         }
